command/ssh-config: option to ignore unresponsive machines
Add option '--ignore-down' to allow ignoring unresponsive machines when
outputting the ssh config settings.

Currently the command will output the ssh-config information for
machines until it reaches a machine with non available and will abort
with an expection.

With multi machine configs it is sometimes reasonable to have machines
down where you wish to retrieve the ssh config information of those
still running. This changes allows you to retrieve the information in a
single run of ssh-config instead of needing to request the information
for each machine individually, which can be expensive depending on the
number of machines and amount of configuration.