communicators/winrm return false on ready? failure
Prior to this patch, the WinRM `.ready?` method raised an error if there was
a communication failure. This behavior differs from the SSH communicator which
returns `false` from `ready?` if a communication failure occurs. The difference
in behavior makes it difficult to write code that works with both types of
communicators.

This patch changes the WinRM `.ready?` method to return false if an error of
type `Vagrant::Errors::VagrantError` occurs, which matches the behavior of the
SSH communicator.

Fixes mitchellh/vagrant#6356.