Much faster rsync auto
This patch aims to make `vagrant rsync-auto` much faster than it used to.

Profiling `rsync-auto`, it was clear that most of the time was spent calling
guest capabilities.

This patch tries to alleviate that issue by refactoring
`VagrantPlugins::SyncedFolderRSync::RsyncHelper` into an instantiable class
(as opposed to only static methods before), so that it can cache a lot of the
results of the calls to the guest capabilities.

We're using that new API in `rsync-auto`, resulting in a dramatic improvement:
on my dev VMs (Mac OS X host, VirtualBox, Ubuntu 14.04 guests), `rsync-auto` now
takes about 300ms per iteration (as opposed to 5-6 secs before); and those 300ms
correspond to almost no overhead compared to running the `rsync` command manually.

The old API with static methods (i.e.
`VagrantPlugins::SyncedFolderRSync::RsyncHelper.single_rsync`) is still there for
backward compatibility reasons, for plugins that might use it (gatling-rsync-auto
is one of these); but a deprecation warning was added to it.

I'm aware this patch will need tests to be merged, but before I spend time writing
those, I'd like to know if this approach is acceptable to the Vagrant core team.