Updated Vagrantfile to install more recent versions of software.
Quieted the apt-get installs to reduce terminal output chaos.

Changed the rvm GPG key server to remove the expired key warning
when the rvm public key is used to verify the rvm download. Command
used conforms to that specified by rvm.io.

Ruby and bundler versions are read from Vagrant's gemspec and used
with rvm to download the latest Ruby version that complies with
the version constraint. Exact bundler version is installed according
to gemspec.

Rubygems is upgraded to latest version through rvm.