add the `salt_call_args` option
This option allows you to pass additional arguments to `salt-call`
during state execution, (similar to `install_args` in installation).
This is useful, for example, for `file-root` and `pillar-root`.

I have also taken the opportunity to reorganize the option string building of
`salt-call` in a similar way to `install`. Instead of multiple methods
that were called once, there is one simple method.