Incorrect length calculation in network_interfaces
It uses `eth_end - 1` rather than `eth_end - eth_start + 1` as the
length paramter for the `slice!` method. In cases when the network
interface list is short, the output is incorrect. For example, if
the array is `['docker0', 'eth0', 'eth1']`, the code will produce
`['eth0', 'docker0', 'eth1']` rather than `['eth0', 'eth1', 'docker0']`.
This PR fixes this.