Update NFS control commands for SUSE
Use systemctl directly on SUSE systems (non-systemd SUSE OSes are
so old that there is no chance of anyone trying to run new vagrant
on them), so that we can pass the --no-pager flag to avoid getting
stuck in a pager while checking NFS server status.

Update both plugin and sudoers script.