Fixed nfs detection on debian Jessie
The recent change to use systemctl to detect and control nfs broke
compatibility with debian Jessie due to the use of the wrong
service name and systemctl list command.

This change reverts the name change and switches to list-units
from list-unit-files as there is no unit file for the service.

The issue was introduced by:
https://github.com/hashicorp/vagrant/pull/8938