Include GitHub CA cert in salt bootstrap script
The downside is that this switches from sh to bash due to array use
and other Bash-isms which might impact *BSD if Bash isn't installed.
Having said that, FreeBSD (at least) seems to need ca_root_nss without
this patch anway so it's just subsituting one package for another in
that specific case.

Fixes #7392