Fix #10604: Network (-ing for Alpine)
Fixes the basic setup of hostname and network on Alpine boxes (...the
ones using ALT Linux as the base plugin. Most of the Alpine boxes being
regularly maintained are using the ALT Linux plugin base. Though the
setup of hostname and networking are broken. This is a minimalistic
fix until there are the proper Alpine plugin modules.